+++ quill/src/handlers/FileHandler.cpp
@@ -7,7 +7,7 @@ namespace
 {
 QUILL_NODISCARD quill::fs::path get_filename(quill::FilenameAppend append_to_filename, quill::fs::path const& filename)
 {
-  if (append_to_filename == quill::FilenameAppend::None)
+  if ((append_to_filename == quill::FilenameAppend::None) || (filename == "/dev/null"))
   {
     return filename;
   }
@@ -33,6 +33,11 @@ FileHandler::FileHandler(fs::path const& filename, std::string const& mode, File
     _fsync(do_fsync)
 {
   open_file(_filename, mode);
+
+  if (_filename == std::string{"/dev/null"})
+  {
+    _is_null = true;
+  }
 }
 
 /***/
@@ -40,6 +45,10 @@ FileHandler::FileHandler(fs::path const& filename, FilenameAppend append_to_file
                          FileEventNotifier file_event_notifier, bool do_fsync)
   : StreamHandler(get_filename(append_to_filename, filename), nullptr, std::move(file_event_notifier)), _fsync(do_fsync)
 {
+  if (_filename == std::string{"/dev/null"})
+  {
+    _is_null = true;
+  }
 }
 
 /***/
@@ -80,6 +89,9 @@ void FileHandler::close_file()
     }
   }
 }
+
+/***/
+bool FileHandler::is_null() const noexcept { return _is_null; }
 
 /***/
 FileHandler::~FileHandler() { close_file(); }